feat: issue-#3 - violation types
- Moved Violation types to 'violation'
  module.
- Replaced all usages of 'ViolationEnum'
  with superseding tyep 'ViolationType'.
