+++ _recycler/inputfilter/scalar.rs
@@ -1,6 +1,6 @@
 use std::fmt::{Debug, Display, Formatter};
 
-use crate::{ViolationEnum, ScalarValue, ViolationTuple, Filter, InputConstraints, Validator, ViolationMessage};
+use crate::{ViolationType, ScalarValue, ViolationTuple, Filter, InputConstraints, Validator, ViolationMessage};
 
 /// Returns a generic message for "Value is missing" violation.
 ///
@@ -57,7 +57,7 @@ impl<'a, T> ScalarInput<'a, T>
     ///
     /// ```rust
     /// use walrs_inputfilter::{
-    ///   ScalarInput, InputConstraints, ViolationEnum,
+    ///   ScalarInput, InputConstraints, ViolationType,
     ///   range_overflow_msg,
     ///   range_underflow_msg,
     ///   scalar_missing_msg_getter,
@@ -96,7 +96,7 @@ impl<'a, T> ScalarInput<'a, T>
         if let Some(min) = self.min {
             if value < min {
                 errs.push((
-                    ViolationEnum::RangeUnderflow,
+                    ViolationType::RangeUnderflow,
                     (self.range_underflow_msg)(self, value),
                 ));
 
@@ -110,7 +110,7 @@ impl<'a, T> ScalarInput<'a, T>
         if let Some(max) = self.max {
             if value > max {
                 errs.push((
-                    ViolationEnum::RangeOverflow,
+                    ViolationType::RangeOverflow,
                     (self.range_overflow_msg)(self, value),
                 ));
 
@@ -189,7 +189,7 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
     ///
     /// ```rust
     /// use walrs_inputfilter::{
-    ///   ScalarInput, InputConstraints, ViolationEnum,
+    ///   ScalarInput, InputConstraints, ViolationType,
     ///   ScalarInputBuilder,
     ///   range_underflow_msg, range_overflow_msg, scalar_missing_msg_getter,
     ///   ScalarValue
@@ -197,7 +197,7 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
     ///
     /// // Setup a custom validator
     /// let validate_is_even = |x: usize| if x % 2 != 0 {
-    ///   Err(vec![(ViolationEnum::CustomError, "Must be even".to_string())])
+    ///   Err(vec![(ViolationType::CustomError, "Must be even".to_string())])
     /// } else {
     ///   Ok(())
     /// };
@@ -258,7 +258,7 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
     ///
     /// ```rust
     /// use walrs_inputfilter::{
-    ///   ScalarInput, InputConstraints, ViolationEnum,
+    ///   ScalarInput, InputConstraints, ViolationType,
     ///   ScalarInputBuilder,
     ///   range_underflow_msg, range_overflow_msg, scalar_missing_msg_getter,
     ///   ScalarValue
@@ -266,7 +266,7 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
     ///
     /// // Setup a custom validator
     /// let validate_is_even = |x: usize| if x % 2 != 0 {
-    ///   Err(vec![(ViolationEnum::CustomError, "Must be even".to_string())])
+    ///   Err(vec![(ViolationType::CustomError, "Must be even".to_string())])
     /// } else {
     ///   Ok(())
     /// };
@@ -288,23 +288,23 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
     ///
     /// let test_cases = [
     ///   ("No value", &usize_required, None, Err(vec![
-    ///     (ViolationEnum::ValueMissing,
+    ///     (ViolationType::ValueMissing,
     ///      scalar_missing_msg_getter(&usize_required)),
     ///   ])),
     ///   ("With valid value", &usize_required, Some(4), Ok(())),
     ///   ("With \"out of lower bounds\" value", &usize_required, Some(0), Err(vec![
-    ///     (ViolationEnum::RangeUnderflow,
+    ///     (ViolationType::RangeUnderflow,
     ///      range_underflow_msg(&usize_required, 0)),
     ///   ])),
     ///   ("With \"out of upper bounds\" value", &usize_required, Some(11), Err(vec![
-    ///     (ViolationEnum::RangeOverflow, range_overflow_msg(&usize_required, 11)),
-    ///     (ViolationEnum::CustomError, "Must be even".to_string()),
+    ///     (ViolationType::RangeOverflow, range_overflow_msg(&usize_required, 11)),
+    ///     (ViolationType::CustomError, "Must be even".to_string()),
     ///   ])),
     ///   ("With \"out of upper bounds\" value, and 'break_on_failure: true'", &usize_break_on_failure, Some(11), Err(vec![
-    ///     (ViolationEnum::RangeOverflow, range_overflow_msg(&usize_required, 11)),
+    ///     (ViolationType::RangeOverflow, range_overflow_msg(&usize_required, 11)),
     ///   ])),
     ///   ("With \"not Even\" value", &usize_required, Some(7), Err(vec![
-    ///     (ViolationEnum::CustomError,
+    ///     (ViolationType::CustomError,
     ///      "Must be even".to_string()),
     ///   ])),
     /// ];
@@ -322,7 +322,7 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
             None => {
                 if self.required {
                     Err(vec![(
-                        ViolationEnum::ValueMissing,
+                        ViolationType::ValueMissing,
                         (self.scalar_missing_msg_getter)(self),
                     )])
                 } else {
@@ -391,7 +391,7 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
     ///   ViolationMessage,
     ///   ViolationTuple,
     ///   InputConstraints,
-    ///   ViolationEnum::CustomError,
+    ///   ViolationType::CustomError,
     ///   scalar_missing_msg_getter
     /// };
     ///
@@ -460,8 +460,8 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
     ///   ScalarInput,
     ///   InputConstraints,
     ///   ViolationMessage,
-    ///   ViolationEnum,
-    ///   ViolationEnum::{
+    ///   ViolationType,
+    ///   ViolationType::{
     ///     CustomError,
     ///     RangeOverflow,
     ///     RangeUnderflow,
@@ -494,7 +494,7 @@ impl<'a, 'b, T> InputConstraints<T, T> for ScalarInput<'a, T>
     /// type TestName = &'static str;
     /// type ConstraintStruct = ScalarInput<'static, usize>;
     /// type TestValue = Option<usize>;
-    /// type ExpectedResult = Result<Option<usize>, Vec<(ViolationEnum, ViolationMessage)>>;
+    /// type ExpectedResult = Result<Option<usize>, Vec<(ViolationType, ViolationMessage)>>;
     ///
     /// let test_cases: Vec<(TestName, &ConstraintStruct, TestValue, ExpectedResult)> = vec![
     ///   ("No value", &usize_input, None, Err(vec![
@@ -579,7 +579,7 @@ impl<T: ScalarValue> Default for ScalarInput<'_, T> {
     ///
     /// ```rust
     /// use walrs_inputfilter::{
-    ///   ScalarInput, InputConstraints, ViolationEnum,
+    ///   ScalarInput, InputConstraints, ViolationType,
     ///   range_overflow_msg, range_underflow_msg, scalar_missing_msg_getter,
     /// };
     ///
@@ -636,7 +636,7 @@ mod test {
     fn test_validate() {
         // Setup a custom validator
         let validate_is_even = |x: usize| if x % 2 != 0 {
-            Err(vec![(ViolationEnum::CustomError, "Must be even".to_string())])
+            Err(vec![(ViolationType::CustomError, "Must be even".to_string())])
         } else {
             Ok(())
         };
@@ -690,7 +690,7 @@ mod test {
         // 1) Test method logic
         // ----
         let validate_is_even = |x: usize| if x % 2 != 0 {
-            Err(vec![(ViolationEnum::CustomError, "Must be even".to_string())])
+            Err(vec![(ViolationType::CustomError, "Must be even".to_string())])
         } else {
             Ok(())
         };
@@ -728,15 +728,15 @@ mod test {
             ("1-10, Even, with valid value", &usize_not_required, Some(2), Ok(())),
             ("1-10, Even, with valid value (2)", &usize_not_required, Some(10), Ok(())),
             ("1-10, Even, with invalid value", &usize_not_required, Some(0), Err(vec![
-                (ViolationEnum::RangeUnderflow,
+                (ViolationType::RangeUnderflow,
                  range_underflow_msg(&usize_not_required, 0))
             ])),
             ("1-10, Even, with invalid value(2)", &usize_not_required, Some(11), Err(vec![
-                (ViolationEnum::RangeOverflow, range_overflow_msg(&usize_not_required, 11)),
-                (ViolationEnum::CustomError, "Must be even".to_string()),
+                (ViolationType::RangeOverflow, range_overflow_msg(&usize_not_required, 11)),
+                (ViolationType::CustomError, "Must be even".to_string()),
             ])),
             ("1-10, Even, with invalid value (3)", &usize_not_required, Some(7), Err(vec![
-                (ViolationEnum::CustomError,
+                (ViolationType::CustomError,
                  "Must be even".to_string()),
             ])),
             ("1-10, Even, with valid value", &usize_not_required, Some(8), Ok(())),
@@ -744,7 +744,7 @@ mod test {
             // Required
             // ----
             ("1-10, Even, required, no value", &usize_required, None, Err(vec![
-                (ViolationEnum::ValueMissing,
+                (ViolationType::ValueMissing,
                  scalar_missing_msg_getter(&usize_required)),
             ])),
             ("1-10, Even, required, with valid value", &usize_required, Some(2), Ok(())),
@@ -752,25 +752,25 @@ mod test {
             ("1-10, Even, required, with valid value (2)", &usize_required, Some(8), Ok(())),
             ("1-10, Even, required, with valid value (3)", &usize_required, Some(10), Ok(())),
             ("1-10, Even, required, with invalid value", &usize_required, Some(0), Err(vec![
-                (ViolationEnum::RangeUnderflow,
+                (ViolationType::RangeUnderflow,
                  range_underflow_msg(&usize_required, 0)),
             ])),
             ("1-10, Even, required, with invalid value(2)", &usize_required, Some(11), Err(vec![
-                (ViolationEnum::RangeOverflow, range_overflow_msg(&usize_required, 11)),
-                (ViolationEnum::CustomError, "Must be even".to_string()),
+                (ViolationType::RangeOverflow, range_overflow_msg(&usize_required, 11)),
+                (ViolationType::CustomError, "Must be even".to_string()),
             ])),
             ("1-10, Even, required, with invalid value (3)", &usize_required, Some(7), Err(vec![
-                (ViolationEnum::CustomError,
+                (ViolationType::CustomError,
                  "Must be even".to_string()),
             ])),
             ("1-10, Even, required, 'break-on-failure: true', with multiple violations", &usize_break_on_failure, Some(11), Err(vec![
-                (ViolationEnum::RangeOverflow, range_overflow_msg(&usize_break_on_failure, 11)),
+                (ViolationType::RangeOverflow, range_overflow_msg(&usize_break_on_failure, 11)),
             ])),
             ("1-10, Even, required, 'break-on-failure: true', with multiple violations (2)", &usize_break_on_failure, Some(0), Err(vec![
-                (ViolationEnum::RangeUnderflow, range_underflow_msg(&usize_break_on_failure, 0)),
+                (ViolationType::RangeUnderflow, range_underflow_msg(&usize_break_on_failure, 0)),
             ])),
             ("1-10, Even, required, with invalid value (3)", &usize_break_on_failure, Some(7), Err(vec![
-                (ViolationEnum::CustomError,
+                (ViolationType::CustomError,
                  "Must be even".to_string()),
             ])),
         ];
@@ -789,7 +789,7 @@ mod test {
             .min(1.0)
             .max(10.0)
             .validators(vec![&|x: f64| if x % 2.0 != 0.0 {
-                Err(vec![(ViolationEnum::CustomError, "Must be even".to_string())])
+                Err(vec![(ViolationType::CustomError, "Must be even".to_string())])
             } else {
                 Ok(())
             }])
@@ -797,13 +797,13 @@ mod test {
             .unwrap();
 
         assert_eq!(f64_input_required.validate_detailed(None), Err(vec![
-            (ViolationEnum::ValueMissing,
+            (ViolationType::ValueMissing,
              scalar_missing_msg_getter(&f64_input_required)),
         ]));
         assert_eq!(f64_input_required.validate_detailed(Some(2.0)), Ok(()));
         assert_eq!(f64_input_required.validate_detailed(Some(11.0)), Err(vec![
-            (ViolationEnum::RangeOverflow, range_overflow_msg(&f64_input_required, 11.0)),
-            (ViolationEnum::CustomError, "Must be even".to_string()),
+            (ViolationType::RangeOverflow, range_overflow_msg(&f64_input_required, 11.0)),
+            (ViolationType::CustomError, "Must be even".to_string()),
         ]));
 
         // Test `char` usage
@@ -817,7 +817,7 @@ mod test {
         assert_eq!(char_input.validate_detailed(Some('a')), Ok(()));
         assert_eq!(char_input.validate_detailed(Some('f')), Ok(()));
         assert_eq!(char_input.validate_detailed(Some('g')), Err(vec![
-            (ViolationEnum::RangeOverflow,
+            (ViolationType::RangeOverflow,
              "`g` is greater than maximum `f`.".to_string()),
         ]));
     }
