+++ walrs_inputfilter/src/lib.rs
@@ -11,9 +11,11 @@ pub mod input;
 pub mod ref_input;
 pub mod traits;
 pub mod validators;
+pub mod violation;
 
 pub use filters::*;
 pub use input::*;
 pub use ref_input::*;
 pub use traits::*;
 pub use validators::*;
+pub use violation::*;
